rvae added

+++ main_rvae.py
@@ -51,7 +51,7 @@ parser.add_argument('--pretrain', default='True', type=str, metavar='N', help='I
 parser.add_argument('-e', '--evaluate', dest='evaluate', action='store_true', help='evaluate model on validation set')
 parser.add_argument('--latent_size', default=50, type=int, metavar='N', help='Size of latent features')
 parser.add_argument('--use_mse', dest='use_mse', action='store_true', help='mse(true) or cross entropy (false)')
-parser.add_argument('--regularizer_gaussian', default=0, type=int, metavar='N', help='regularizer_gaussian')
+parser.add_argument('--regularizer_gaussian', default=0, type=float, metavar='N', help='regularizer_gaussian')
 parser.add_argument('--gaussian_target_distance', default=5, type=float, metavar='N', help='gaussian_target_distance')
